Desktop/report-api-v2/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
Deprecated==1.2.14
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
django-oauth-toolkit==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
jwcrypto==1.5.0
Markdown==3.5
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
mysql-connector-repackaged==0.3.1
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.2
protobuf==4.21.12
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
whitenoise==6.6.0
wrapt==1.15.0
zipp==3.17.0
pre-commit==3.5.0
ruff==0.1.6
django-environ==0.11.2
